Add a camera 'buffer' range, which allows the camera to view a region normally outside of the strict world range.
Updates level building to always add a border of type EDGE around the map, so ships can't just fly off into oblivion.

Added a tile type of EDGE which just looks like a solid block - not sure if this should be the final texture. 

Centered the camera on game load.